requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
groq==0.4.2
pydantic==2.5.3
sqlalchemy==2.0.25
pymysql==1.1.0
